+++ src/fft/fft_driver.cpp
@@ -84,9 +84,9 @@ FFTDriver::FFTDriver(Mesh *pm, ParameterInput *pin) {
     lx3max = lx3max>lx3?lx3min:lx3;
   }
 
-  int nbx1=lx1max-lx1min+1;
-  int nbx2=lx2max-lx2min+1;
-  int nbx3=lx3max-lx3min+1;
+  int nbx1 = static_cast<int>(lx1max-lx1min+1);
+  int nbx2 = static_cast<int>(lx2max-lx2min+1);
+  int nbx3 = static_cast<int>(lx3max-lx3min+1);
 
   nmb = nbx1*nbx2*nbx3; // number of mesh blocks to be loaded to the FFT block
   if (pm->nbtotal/nmb != nranks_) {
@@ -111,9 +111,9 @@ FFTDriver::FFTDriver(Mesh *pm, ParameterInput *pin) {
     fft_loclist_[n].lx2 = fft_loclist_[n].lx2/nbx2;
     fft_loclist_[n].lx3 = fft_loclist_[n].lx3/nbx3;
   }
-  npx1=pm->nrbx1/nbx1;
-  npx2=pm->nrbx2/nbx2;
-  npx3=pm->nrbx3/nbx3;
+  npx1 = static_cast<int>(pm->nrbx1/nbx1);
+  npx2 = static_cast<int>(pm->nrbx2/nbx2);
+  npx3 = static_cast<int>(pm->nrbx3/nbx3);
 
   fft_mesh_size_=pm->mesh_size;
